hopefully

+++ src/Scbwi/Models/Package.cs
@@ -9,8 +9,8 @@ namespace Scbwi.Models {
 
         public string title { get; set; }
         public string description { get; set; }
-        public double regularprice { get; set; }
-        public double lateprice { get; set; }
+        public decimal regularprice { get; set; }
+        public decimal lateprice { get; set; }
         public bool member { get; set; }
         public int max { get; set; }
     }
